Draper changes 2 (#14)
* Linking against VmfTestLib

removed the install for yaml-cpp in the dockerfile and pointed cmake at vmf's yaml-cpp instead

* Changed rand argument types from size_t to ul

Arguments to vmf::VmfRand::randBetween are now casted to (or declared as) unsigned long before being passed. The result is casted back to size_t to minimize codebase changes.

* Remove legacy AFL++ mutators and update build

Remove many AFLPlusPlus mutator source and header files (various AFL*Mutator.cpp/.hpp) and stop including them in the AFLPlusPlus library. Update AFLPlusPlus/CMakeLists.txt to remove the deleted sources and link against vmf_core_modules. Adjust AFLDeleteMutator.hpp: simplify includes, change choose_block_len signature to use VmfRand& and update the rand member. Also update top-level CMakeLists and test/config/experimentalModules.yaml to reflect these module removals.

* Added ASAN options

* Fixed test-related memory bugs

* fixed OoB index calc in AFLInteresting32Mutator

The interesting_32 array is composed of 28 int64_t elements (9 from INTERESTING_8, 10 from INTERESTING_16, 9 from INTERESTING_32). The random index into this array was computed using sizeof() >> 2 (divide by 4), which is correct for a 32-bit element type but wrong for int64_t (8 bytes). This produced a range of 0..55 against an array of only 28 elements, causing an out-of-bounds read detected by ASAN as a global-buffer-overflow. Replaced the sizeof() bit-shift with the explicit sum of the INTERESTING_8_LEN, INTERESTING_16_LEN, and INTERESTING_32_LEN constants defined in config.h.

* draper patches

* Fixed failure case in WidenCodePointMutator

Replace the repeated random-try loop with a scan that collects all printable ASCII byte indices, then picks one at random. If no printable bytes are found, copy the buffer unchanged and return. This avoids repeated attempts/infinite loops and adds a fallback case.

+++ Radamsa/vmf/src/modules/common/mutator/RadamsaInsertByteMutator.cpp
@@ -141,16 +141,16 @@ void RadamsaInsertByteMutator::mutateTestCase(StorageModule& storage, StorageEnt
 
     // Select a random index from which the new byte will be inserted.
 
-    const size_t lower{0u};
+    const unsigned long lower{0ul};
     const size_t upper{originalSize - 1u};
-    const size_t maximumRandomIndexValue{originalSize - minimumSeedIndex};
+    const unsigned long maximumRandomIndexValue{static_cast<unsigned long>(originalSize - minimumSeedIndex)};
     const size_t randomInsertionIndex{
                                 std::clamp(
-                                    rand->randBetween(
+                                    static_cast<size_t>(rand->randBetween(
                                         lower,
                                         maximumRandomIndexValue
-                                    ) + minimumSeedIndex,
-                                    lower,
+                                    )) + minimumSeedIndex,
+                                    static_cast<size_t>(lower),
                                     upper
                                 )
     };
@@ -163,7 +163,7 @@ void RadamsaInsertByteMutator::mutateTestCase(StorageModule& storage, StorageEnt
         newBuffer[destinationIndex] = originalBuffer[sourceIndex];
 
         if (sourceIndex == randomInsertionIndex)
-            newBuffer[++destinationIndex] = static_cast<char>(rand->randBetween(0u, std::numeric_limits<char>::max()));
+            newBuffer[++destinationIndex] = static_cast<char>(rand->randBetween(0ul, static_cast<unsigned long>(std::numeric_limits<char>::max())));
 
         ++destinationIndex;
     }
